# services/gateway/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_fieldstone_fieldstone_services_requests_handlers.listResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      requests:
        items:
          $ref: '#/definitions/handlers.ServiceRequestResponse'
        type: array
      total:
        type: integer
    type: object
  github_com_fieldstone_fieldstone_services_requests_handlers.updateStatusRequest:
    properties:
      role:
        type: string
      status:
        type: string
    type: object
...
  handlers.ServiceRequestResponse:
    properties:
      assigned_to:
        type: string
      closed_at:
        type: string
      created_at:
        type: string
      department_id:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        items:
          type: integer
        type: array
      metadata:
        items:
          type: integer
        type: array
      request_type:
        type: string
      status:
        type: string
      submitter_email:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.assignRequest:
    properties:
      assigned_to:
        type: string
      role:
        type: string
    type: object
...
  handlers.createServiceRequestRequest:
    properties:
      department_id:
        type: string
      description:
        type: string
      location:
        items:
          type: integer
        type: array
      metadata:
        items:
          type: integer
        type: array
      request_type:
        type: string
      submitter_email:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Fieldstone Contributors
    url: https://github.com/fieldstone/fieldstone
  description: |-
    A self-hostable platform for city governments to run common civic services.

    Staff endpoints require a Bearer JWT from the configured OIDC provider.
    Citizen endpoints (marked "public") do not require authentication.

    **Development**: set `DEV_DISABLE_AUTH=true` in .env to bypass JWT
    validation locally. Remove before production deployment.
  license:
    name: MIT
    url: https://github.com/fieldstone/fieldstone/blob/main/LICENSE
  title: Fieldstone Civic Services API
  version: 0.1.0
paths:
...
  /v1/requests:
    get:
      parameters:
      - default: 20
        description: Max results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Filter by status (open, assigned, in_progress, resolved, closed)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fieldstone_fieldstone_services_requests_handlers.listResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List service requests (staff)
      tags:
      - requests
    post:
      consumes:
      - application/json
      description: |-
        Public endpoint — no authentication required. Citizens use this to
        report issues such as potholes, broken streetlights, or code violations.
      parameters:
      - description: Service request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createServiceRequestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ServiceRequestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Metadata validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit a 311 service request (public)
      tags:
      - requests
  /v1/requests/{id}:
    get:
      parameters:
      - description: Request UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ServiceRequestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a service request
      tags:
      - requests
  /v1/requests/{id}/assign:
    patch:
      consumes:
      - application/json
      description: |-
        Validates the open→assigned workflow transition. Updates assigned_to
        and status atomically. Publishes service_request.assigned.
      parameters:
      - description: Request UUID
        in: path
        name: id
        required: true
        type: string
      - description: Assignment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.assignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ServiceRequestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Transition not allowed by workflow
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign a request to a staff member
      tags:
      - requests
  /v1/requests/{id}/status:
    patch:
      consumes:
      - application/json
      description: |-
        Validates the transition via the workflow service. Terminal statuses
        (resolved, closed) also set closed_at.
      parameters:
      - description: Request UUID
        in: path
        name: id
        required: true
        type: string
      - description: Status transition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fieldstone_fieldstone_services_requests_handlers.updateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ServiceRequestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Transition not allowed by workflow
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update service request status
      tags:
      - requests
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"